+++ src/main/kotlin/org/jetbrains/research/testspark/tools/GenerationTool.kt
@@ -1,0 +1,14 @@
+package org.jetbrains.research.testspark.tools
+
+/**
+ * Enum class representing strategy for test generation
+ */
+enum class GenerationTool(val toolId: String) {
+    EvoSuite("EvoSuite"),
+    LLM("LLM"),
+    ;
+
+    companion object {
+        fun from(findValue: String): GenerationTool = entries.first { it.toolId == findValue }
+    }
+}