+++ src/main/kotlin/org/jetbrains/research/testspark/tools/evosuite/EvoSuite.kt
@@ -184,6 +184,7 @@ class EvoSuite(override val name: String = "EvoSuite") : Tool {
             testGenerationController,
             testSparkDisplayManager,
             testsExecutionResultManager,
+            name,
         )
     }
 }