+++ CHANGELOG.md
@@ -4,6 +4,15 @@
 
 ## Unreleased
 
+## 0.3.1
+
+### Fixed
+- Multiple minor bugs in test generation algorithms.
+
+### Added
+- Kotlin test generation now works in K2 mode as well.
+- Support IDEA `243.*`.
+
 ## 0.3.0
 
 ### Fixed